An improved sumOfSquares() function.
The existing code includes a function named foo() that computes
the sum of the squares of the numbers in an array. This patch
renames that function to sumOfSquares(), adds documentation and
tests for it, and re-implements it to add type checking, to
avoid allocation of an intermediate array, and to make it work
for any iterable argument instead of just arrays.

Note that this re-implementation causes breaking changes to the API:

1) With no argument, `foo()` returned 0, but `sumOfSquares()`
   throws TypeError instead.
2) JavaScript type conversion could cause unexpected return values
   when foo() was invoked with an array of non-numeric values. For
   example `foo([true,true])` returns 2, and `foo([[]])` returns 0.
   When sumOfSquares is invoked with those values, it returns NaN.