Fix the openUrlsInOrder() function.
This patch fixes the openUrlsInOrder() function so that its
tests will pass. The problem is that the previous code was
repeatedly invoking an async function with a forEach() loop
without awaiting the return value of each invocation. This
patch switches to a `for/of` loop which allows us to remove
the extra async function so that we don't need to `await` it.